requirements/requirements.txt
django==4.2.7
psycopg2-binary==2.9.6
gunicorn==20.1.0
python-dotenv==1.0.0
django-environ
